* Feature: Added device tracker * Refactor : coordinator in its own file (allowing more and better HA entities)
