+++ custom_components/abetterrouteplanner/strings.json
@@ -70,6 +70,11 @@
       "calib_ref_cons": {
         "name":"Calibration Reference Consumption"
       }
+    },
+    "device_tracker": {
+      "location": {
+        "name": "Location"
+      }
     }
   }
 }
