+++ custom_components/abetterrouteplanner/translations/de.json
@@ -70,6 +70,11 @@
       "calib_ref_cons": {
         "name": "Kalibrierter Referenzverbrauch"
       }
+    },
+    "device_tracker": {
+      "location": {
+        "name": "Standort"
+      }
     }
   }
 }
