+++ custom_components/abetterrouteplanner/translations/sv.json
@@ -70,6 +70,11 @@
       "calib_ref_cons": {
         "name": "Kalibrerad referensförbrukning"
       }
+    },
+    "device_tracker": {
+      "location": {
+        "name": "Plats"
+      }
     }
   }
 }
